pyproject.toml
[project]
name = "find-me-nest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dspy-ai>=2.6.23",
    "googlemaps>=4.10.0",
    "gspread>=6.2.0",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.8",
]
